move SImageDescription to types
